worker/requirements.txt
aio-pika==9.4.3
motor==3.7.0
pymongo==4.10.1
httpx==0.27.0
boto3==1.34.0
rasterio==1.3.9
numpy==1.26.4
pillow==10.3.0
python-dotenv==1.0.0
